Move export related logic into an export directory
