+++ src/originalBudgetTrackingApp/export/outputVendors/ynab/ynab.ts
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import * as ynab from 'ynab';
 import moment from 'moment/moment';
 import { EnrichedTransaction, OutputVendor, OutputVendorName } from '@/originalBudgetTrackingApp/commonTypes';
-import { Config, getConfig, YnabConfig } from '../../configManager/configManager';
+import { Config, getConfig, YnabConfig } from '../../../configManager/configManager';
 
 const INITIAL_YNAB_ACCESS_TOKEN = 'AABB';
 const YNAB_DATE_FORMAT = 'YYYY-MM-DD';
